requirement.txt
numpy==1.26.0
pandas==2.1.4
pillow==9.0.1
torch==2.1.1+cu118
transformers==4.34.0
scikit-learn==1.0.2